requirements.txt
Django==5.0.6
gunicorn==21.2.0
mysqlclient==2.2.4
djangorestframework==3.15.1
django-cors-headers==4.4.0